+++ index.js
@@ -1,10 +1,8 @@
-const wmts = require('./src/wmts')
-const service = require('./src/service')
-
 /**
  * OGC Parser
  */
 module.exports = {
-  service: service,
-  wmts: wmts
+  service: require('./src/service'),
+  wmts: require('./src/wmts'),
+  wms: require('./src/wms')
 }