+++ test/xml.js
@@ -1,24 +1,32 @@
 const path = require('path')
 const fs = require('fs')
 
-const folder = (filename) => path.join(__dirname, 'in', filename)
+const wms = (filename) => path.join(__dirname, 'in', 'wms', filename)
+const wmts = (filename) => path.join(__dirname, 'in', 'wmts', filename)
 
 module.exports = {
   arcgis: {
-    wmts: fs.readFileSync(folder('arcgis-wmts.xml'), 'utf8'),
-    wmtsError: fs.readFileSync(folder('arcgis-wmts-error.xml'), 'utf8')
+    wmts: fs.readFileSync(wmts('arcgis-wmts.xml'), 'utf8'),
+    wmtsError: fs.readFileSync(wmts('arcgis-wmts-error.xml'), 'utf8')
   },
   geoserver: {
-    wmts: fs.readFileSync(folder('geoserver-wmts.xml'), 'utf8')
+    wmts: fs.readFileSync(wmts('geoserver-wmts.xml'), 'utf8')
   },
   mapbox: {
-    wmts: fs.readFileSync(folder('mapbox-wmts.xml'), 'utf8')
-  },
-  toporama: {
-    wms: fs.readFileSync(folder('toporama-wms.xml'), 'utf8')
+    wmts: fs.readFileSync(wmts('mapbox-wmts.xml'), 'utf8')
   },
   mapProxy: {
-    wmtsKvp: fs.readFileSync(folder('mapProxy-wmtsKvp.xml'), 'utf8'),
-    wmtsRestful: fs.readFileSync(folder('mapProxy-wmtsRestful.xml'), 'utf8')
+    wmtsKvp: fs.readFileSync(wmts('map-proxy-wmts-kvp.xml'), 'utf8'),
+    wmtsRestful: fs.readFileSync(wmts('map-proxy-wmts-restful.xml'), 'utf8')
+  },
+  toporama: {
+    wms: fs.readFileSync(wms('toporama-wms.xml'), 'utf8')
+  },
+  ogc: {
+    wms100: fs.readFileSync(wms('ogc-wms-1.0.0.xml'), 'utf8'),
+    wms110: fs.readFileSync(wms('ogc-wms-1.1.0.xml'), 'utf8'),
+    wms111: fs.readFileSync(wms('ogc-wms-1.1.1.xml'), 'utf8'),
+    wms130: fs.readFileSync(wms('ogc-wms-1.3.0.xml'), 'utf8'),
+    wms130Exception: fs.readFileSync(wms('ogc-wms-1.3.0-exception.xml'), 'utf8')
   }
 }